requirements.txt
python-telegram-bot==20.3
yfinance
numpy
groq